Upgrades round to int dmg nums

+++ TacticTowers/Assets/Prefabs/Upgrades/Damage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Damage : Upgrade
 {
-    [SerializeField] private int bonus;
+    [SerializeField] private float bonus;
     
     public override void Execute(Tower tower)
     {
